% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake-package.R
\docType{package}
\name{rake-package}
\alias{rake}
\alias{rake-package}
\title{rake: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
